Test if /etc/grub.d/20_linux_xen is there

+++ wheezy2jessie.sh
@@ -81,8 +81,8 @@ sed -i "s/^ rocommunity6 public/# rocommunity6 public/" $CFG
 sed -i "s/agentAddress  udp:127/#agentAddress  udp:127/" $CFG
 
 # fix our xen modification
-rm -rf /etc/grub.d/09_linux_xen
-dpkg-divert --divert /etc/grub.d/09_linux_xen --rename /etc/grub.d/20_linux_xen
+[ -f /etc/grub.d/20_linux_xen ] && rm -rf /etc/grub.d/09_linux_xen && \
+ dpkg-divert --divert /etc/grub.d/09_linux_xen --rename /etc/grub.d/20_linux_xen
 
 # chrony update
 if [ -f /etc/chrony/chrony.conf.new ]; then CFG=/etc/chrony/chrony.conf.new; else CFG=/etc/chrony/chrony.conf; fi
